// src/lib/api/models/createLeadProtectionPolicyDto.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * RealHub API
 * Multi-tenant Real Estate Platform API
 * OpenAPI spec version: v1
 */
import type { CreateLeadProtectionPolicyDtoSource } from './createLeadProtectionPolicyDtoSource';
import type { CreateLeadProtectionPolicyDtoSellingMode } from './createLeadProtectionPolicyDtoSellingMode';
import type { CreateLeadProtectionPolicyDtoCustomerType } from './createLeadProtectionPolicyDtoCustomerType';

export interface CreateLeadProtectionPolicyDto {
  name: string;
  source?: CreateLeadProtectionPolicyDtoSource;
  propertyTypeId?: string;
  sellingMode?: CreateLeadProtectionPolicyDtoSellingMode;
  customerType?: CreateLeadProtectionPolicyDtoCustomerType;
  protectionDays?: number;
  inactiveReclaimDays?: number;
  allowReassign?: boolean;
  conditionJson?: string;
  priority?: number;
}