// src/lib/api/models/updateLeadProtectionPolicyDto.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * RealHub API
 * Multi-tenant Real Estate Platform API
 * OpenAPI spec version: v1
 */

export interface UpdateLeadProtectionPolicyDto {
  name?: string;
  protectionDays?: number;
  inactiveReclaimDays?: number;
  allowReassign?: boolean;
  priority?: number;
}